implement loading

+++ src/store/actions/auth.js
@@ -1,38 +1,35 @@
-import auth from '../../services/auth';
-import { alertError, alertSuccess } from './alert';
+import auth from '../../services/auth'
+import { alertError, alertSuccess } from './alert'
 
 export const changeValue = e => {
-    const { name, value } = e.target;
+    const { name, value } = e.target
     return { type : 'VALUE_CHANGED_LOGIN', payload : { name , data : value } }
 }
 
 export const login = ( email, password ) => async dispatch => {
-    try{
-        const { data } = await auth.login(email, password)
-        if(!data.user.blocked){
-            localStorage.setItem('jwt', data.jwt)
-            localStorage.setItem('user', data.user.id)
-            localStorage.setItem('name', data.user.username)
-            localStorage.setItem('email', data.user.email)
-            localStorage.setItem('role', data.user.role.type);
-            dispatch({ type : 'LOGIN'})
-        } else {
-            dispatch(alertError('Usuário bloqueado'))
-        }
-    } catch ({ response }) {
-        dispatch(alertError(response.data.message))
+    const { data, error } = await auth.login(email, password, dispatch)
+    if (error) {
+        dispatch(alertError(error))
+    } else if(!data.user.blocked && data.user.confirmed){
+        localStorage.setItem('jwt', data.jwt)
+        localStorage.setItem('user', data.user.id)
+        localStorage.setItem('name', data.user.username)
+        localStorage.setItem('email', data.user.email)
+        localStorage.setItem('role', data.user.role.type)
+        dispatch({ type : 'LOGIN'})
+    } else {
+        dispatch(alertError('Usuário bloqueado'))
     }
 }
 
 export const singup = ( username, email, password ) => async dispatch => {
-    try {
-        const response = await auth.register( username, email, password)
-        localStorage.setItem('jwt', response.data.jwt);
+    const { data, error } = await auth.register( username, email, password, dispatch)
+    if (error){
+        dispatch(alertError(error))
+    } else {
+        localStorage.setItem('jwt', data.jwt)
         dispatch(alertSuccess('Cadastro efetuado com sucesso, verifique o email!'))
-    } catch ({ response }) {
-        dispatch(alertError(response.data.message))
     }
-    
 }   
 
 export const logout = () => dispatch => {
@@ -41,21 +38,22 @@ export const logout = () => dispatch => {
 }
 
 export const forgotte_password = email => async dispatch => {
-    try {
-        const response = await auth.forgotte_password(email)
-        if (response.data.ok)
+    const { data, error } = await auth.forgotte_password(email, dispatch)
+    if (error){
+        dispatch(alertError(error))
+    } else if (data.ok){        
             dispatch(alertSuccess('Email enviado'))
-    } catch ({ response }) {
-        dispatch(alertError(response.data.message))
+    } else {
+        dispatch(alertSuccess('Erro desconhecido'))
     }
 }
 
 export const reset_password = (code, password, passwordConfirmation ) => async dispatch => {
-    try {
-        const response = await auth.reset_password(code, password, passwordConfirmation)
-        localStorage.setItem('jwt', response.data.jwt);
+    const { data, error } = await auth.reset_password(code, password, passwordConfirmation, dispatch)
+    if (error) {
+        dispatch(alertError(error))
+    } else {
+        localStorage.setItem('jwt', data.jwt)
         dispatch({ type : 'LOGIN'})
-    } catch ({ response }) {
-        dispatch(alertError(response.data.message))
     }
 }
